Send EVENT callbacks immediately rather than send to main task queue for later processing.
Callbacks were being handled asynchronously... in certain circumstances this caused data to change from when the callback EVENT was requested until it was actually sent.
This update changes all queue_event() calls into a send_event() but ONLY when the request originates from the main task thread. If the event is triggered within a timer or a different task (e.g. sync task) then to ensure that callbacks are always from the same task thread, those continue to use queue_event() so we don't have to worry about user callback code being re-entrant.